standard views shift modifier now selects oposite direction ctrl modifier directs it at maingl 2

+++ src/gui/widgets/glwidget.cpp
@@ -228,7 +228,7 @@ void GLWidget::wheelEvent( QWheelEvent *event )
 
 void GLWidget::setView( Fn::Orient view )
 {
-    m_arcBall->setView( (int)view );
+    m_arcBall->setView( view );
     update();
 }
